+++ app/views/albums/_form.html.erb
@@ -20,7 +20,7 @@
   </div>
   <%= form.date_field :released_on, max: Time.zone.today, class: 'w-full mb-3' %>
   <%= form.labels :license, text: "License" %>
-  <%= form.collection_select :license_id, License.all, :id, :label, include_blank: "Select a license" %>
+  <%= form.collection_select :license_id, License.all, :id, :label %>
   <section id="tracks">
     <h2 class="text-2xl tracking-tight leading-10 font-extrabold my-3">Tracks</h2>
     <div class="flex items-center gap-2 mb-4">
