+++ db/migrate/20240130201122_add_license_to_album.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class AddLicenseToAlbum < ActiveRecord::Migration[7.1]
   disable_ddl_transaction!
